+++ example/lib/main.dart
@@ -23,11 +23,9 @@ class _HomePageState extends State<HomePage> {
     return MaterialApp(
       debugShowCheckedModeBanner: false,
       theme: ThemeData.from(
-          colorScheme: ColorScheme.fromSeed(
-              seedColor: Colors.blue, brightness: Brightness.light)),
+          colorScheme: ColorScheme.fromSeed(seedColor: Colors.blue, brightness: Brightness.light)),
       darkTheme: ThemeData.from(
-          colorScheme: ColorScheme.fromSeed(
-              seedColor: Colors.blue, brightness: Brightness.dark)),
+          colorScheme: ColorScheme.fromSeed(seedColor: Colors.blue, brightness: Brightness.dark)),
       themeMode: themeMode,
       home: ScaffoldMessenger(
         child: Scaffold(
@@ -102,6 +100,7 @@ class _HomeBodyState extends State<HomeBody> {
                 PositionViewPart(),
                 TapOffsetPart(),
                 CheckBoxPart(),
+                TogglePart(),
                 CounterPart(),
                 SelectPart(),
                 FormTextInput(
@@ -175,8 +174,7 @@ class _HomeBodyState extends State<HomeBody> {
                     style: FormPrimaryButton.dangerStyle,
                     child: Text(
                       'Primary Danger',
-                      style:
-                          TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
+                      style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
                     )),
                 FormSecondaryButton(
                     onPressed: () {
@@ -205,8 +203,7 @@ class _HomeBodyState extends State<HomeBody> {
                     style: FormSecondaryButton.dangerStyle,
                     child: Text(
                       'Secondary Danger',
-                      style:
-                          TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
+                      style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
                     )),
               ],
             ),
@@ -289,18 +286,37 @@ class CounterPart extends StatefulWidget {
 
 class _CounterPartState extends State<CounterPart> {
   int value = 0;
+  int valueCustom = 0;
+
   @override
   Widget build(BuildContext context) {
-    return FormCounter(
-      hint: 'Counter',
-      max: 1000,
-      min: 0,
-      value: value,
-      onChange: (v) {
-        setState(() {
-          value = v;
-        });
-      },
+    return Column(
+      spacing: 10,
+      children: [
+        FormCounter(
+          hint: 'Counter',
+          max: 1000,
+          min: 0,
+          value: value,
+          onChange: (v) {
+            setState(() {
+              value = v;
+            });
+          },
+        ),
+        FormCounter(
+          hint: 'Counter (Step 5)',
+          max: 1000,
+          min: 0,
+          step: 5,
+          value: valueCustom,
+          onChange: (v) {
+            setState(() {
+              valueCustom = v;
+            });
+          },
+        ),
+      ],
     );
   }
 }
@@ -367,10 +383,10 @@ class _PositionSelectPartState extends State<PositionSelectPart> {
               return FormPositionSelect(
                 bytes: snapshot.data!.buffer.asUint8List(),
                 allowSelectZone: [
-                  Rect.fromLTRB(43.5374149659864, 11.1975116640747,
-                      46.8537414965986, 88.4914463452566),
-                  Rect.fromLTRB(46.8537414965986, 11.1975116640747,
-                      53.1462585034014, 46.9673405909798),
+                  Rect.fromLTRB(
+                      43.5374149659864, 11.1975116640747, 46.8537414965986, 88.4914463452566),
+                  Rect.fromLTRB(
+                      46.8537414965986, 11.1975116640747, 53.1462585034014, 46.9673405909798),
                   // Rect.fromLTRB(53.1462585034014, 11.1975116640747,
                   //     56.4625850340136, 88.4914463452566),
                   // Rect.fromLTRB(46.8537414965986, 53.343701399689,
@@ -497,5 +513,35 @@ class _TapOffsetPartState extends State<TapOffsetPart> {
 
   Future<ByteData> getBgBytes() async {
     return await rootBundle.load('assets/example.png');
+  }
+}
+
+class TogglePart extends StatefulWidget {
+  const TogglePart({super.key});
+
+  @override
+  State<TogglePart> createState() => _TogglePartState();
+}
+
+class _TogglePartState extends State<TogglePart> {
+  bool value = false;
+  @override
+  Widget build(BuildContext context) {
+    return FormToggle(
+      value: value,
+      hint: 'Toggle Button',
+      onChange: (v) {
+        setState(() {
+          value = v;
+        });
+        ScaffoldMessenger.of(context).showSnackBar(SnackBar(
+          content: Text(
+            'Toggle: ${v ? 'On' : 'Off'}',
+          ),
+          behavior: SnackBarBehavior.floating,
+          duration: Duration(milliseconds: 300),
+        ));
+      },
+    );
   }
 }